# src/main/java/com/btcplatform/trading/model/Transaction.java
package com.btcplatform.trading.model;

import javax.persistence.*;
import java.math.BigDecimal;
import java.time.LocalDateTime;

@Entity
@Table(name = "transactions")
public class Transaction {

    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;

    @ManyToOne
    @JoinColumn(name = "user_id", nullable = false)
    private User user;

    @Column(nullable = false)
    private String type;

    private BigDecimal btcAmount;
    private BigDecimal usdAmount;
    private BigDecimal btcPrice;
    private LocalDateTime timestamp;

    // Getters and Setters
    public Long getId() { return id; }
    public void setId(Long id) { this.id = id; }

    public User getUser() { return user; }
    public void setUser(User user) { this.user = user; }

    public String getType() { return type; }
    public void setType(String type) { this.type = type; }

    public BigDecimal getBtcAmount() { return btcAmount; }
    public void setBtcAmount(BigDecimal btcAmount) { this.btcAmount = btcAmount; }

    public BigDecimal getUsdAmount() { return usdAmount; }
    public void setUsdAmount(BigDecimal usdAmount) { this.usdAmount = usdAmount; }

    public BigDecimal getBtcPrice() { return btcPrice; }
    public void setBtcPrice(BigDecimal btcPrice) { this.btcPrice = btcPrice; }

    public LocalDateTime getTimestamp() { return timestamp; }
    public void setTimestamp(LocalDateTime timestamp) { this.timestamp = timestamp; }
}